Adding classes for handshaking, moving things around
