+++ src/SlackConnector.Tests.Unit/SlackConnector.Tests.Unit.csproj
@@ -79,7 +79,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SlackConnectorTests\ConnectTests.cs" />
     <Compile Include="SlackConnectorTests\OnMessageTests.cs" />
-    <Compile Include="Sockets\Messages\MessageInterpreterTests.cs" />
+    <Compile Include="Connections\Sockets\Messages\MessageInterpreterTests.cs" />
     <Compile Include="Stubs\WebSocketStub.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -92,6 +92,9 @@
     <None Include="app.config" />
     <None Include="packages.config" />
     <EmbeddedResource Include="Resources\HandShakeResponse.json" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Connections\Handshaking\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
