+++ src/SlackConnector/Connections/Handshaking/HandShaker.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+using SlackConnector.Connections.Handshaking.Models;
+
+namespace SlackConnector.Connections.Handshaking
+{
+    internal class HandShaker : IHandshaker
+    {
+        public Task<SlackHandshake> FirmShake(string slackKey)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
