+++ src/SlackConnector/Connections/Handshaking/IHandshaker.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+using SlackConnector.Connections.Handshaking.Models;
+
+namespace SlackConnector.Connections.Handshaking
+{
+    internal interface IHandshaker
+    {
+        /// <summary>
+        /// No one likes a limp shake - AMIRITE?
+        /// </summary>
+        Task<SlackHandshake> FirmShake(string slackKey);
+    }
+}
