+++ src/SlackConnector/Connections/Handshaking/Models/Detail.cs
@@ -1,4 +1,4 @@
-﻿namespace SlackConnector.Models.Handshake
+﻿namespace SlackConnector.Connections.Handshaking.Models
 {
     internal class Detail
     {
