+++ src/SlackConnector/Connections/Handshaking/Models/Im.cs
@@ -1,4 +1,4 @@
-﻿namespace SlackConnector.Models.Handshake
+﻿namespace SlackConnector.Connections.Handshaking.Models
 {
     internal class Im
     {
