+++ src/SlackConnector/Connections/Handshaking/Models/User.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace SlackConnector.Models.Handshake
+namespace SlackConnector.Connections.Handshaking.Models
 {
     internal class User : Detail
     {
