+++ src/SlackConnector/Connections/Sockets/Messages/MessageSubType.cs
@@ -1,4 +1,4 @@
-﻿namespace SlackConnector.Sockets.Messages
+﻿namespace SlackConnector.Connections.Sockets.Messages
 {
     internal enum MessageSubType
     {
