+++ src/SlackConnector/SlackConnector.csproj
@@ -58,16 +58,18 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BotHelpers\BotNameRegexComposer.cs" />
+    <Compile Include="Connections\Handshaking\HandShaker.cs" />
+    <Compile Include="Connections\Handshaking\IHandshaker.cs" />
     <Compile Include="EventHandlers\ConnectionStatusChangedEventHandler.cs" />
     <Compile Include="Models\BotMessage.cs" />
     <Compile Include="EventHandlers\MessageReceivedEventHandler.cs" />
     <Compile Include="ISlackConnector.cs" />
-    <Compile Include="Models\Handshake\Channel.cs" />
-    <Compile Include="Models\Handshake\SlackHandshake.cs" />
-    <Compile Include="Models\Handshake\Detail.cs" />
-    <Compile Include="Models\Handshake\Im.cs" />
-    <Compile Include="Models\Handshake\User.cs" />
-    <Compile Include="Models\Handshake\UserProfile.cs" />
+    <Compile Include="Connections\Handshaking\Models\Channel.cs" />
+    <Compile Include="Connections\Handshaking\Models\SlackHandshake.cs" />
+    <Compile Include="Connections\Handshaking\Models\Detail.cs" />
+    <Compile Include="Connections\Handshaking\Models\Im.cs" />
+    <Compile Include="Connections\Handshaking\Models\User.cs" />
+    <Compile Include="Connections\Handshaking\Models\UserProfile.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Models\ResponseContext.cs" />
     <Compile Include="Models\SlackAttachment.cs" />
@@ -78,15 +80,15 @@
     <Compile Include="SlackConnector.cs" />
     <Compile Include="Models\SlackMessage.cs" />
     <Compile Include="Models\SlackUser.cs" />
-    <Compile Include="Sockets\IWebSocket.cs" />
-    <Compile Include="Sockets\IWebSocketFactory.cs" />
-    <Compile Include="Sockets\Messages\IMessageInterpreter.cs" />
-    <Compile Include="Sockets\Messages\InboundMessage.cs" />
-    <Compile Include="Sockets\Messages\MessageInterpreter.cs" />
-    <Compile Include="Sockets\Messages\MessageSubType.cs" />
-    <Compile Include="Sockets\Messages\MessageType.cs" />
-    <Compile Include="Sockets\WebSocket.cs" />
-    <Compile Include="Sockets\WebSocketFactory.cs" />
+    <Compile Include="Connections\Sockets\IWebSocket.cs" />
+    <Compile Include="Connections\Sockets\IWebSocketFactory.cs" />
+    <Compile Include="Connections\Sockets\Messages\IMessageInterpreter.cs" />
+    <Compile Include="Connections\Sockets\Messages\InboundMessage.cs" />
+    <Compile Include="Connections\Sockets\Messages\MessageInterpreter.cs" />
+    <Compile Include="Connections\Sockets\Messages\MessageSubType.cs" />
+    <Compile Include="Connections\Sockets\Messages\MessageType.cs" />
+    <Compile Include="Connections\Sockets\WebSocket.cs" />
+    <Compile Include="Connections\Sockets\WebSocketFactory.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
